+++ src/game/movement.c
@@ -30,15 +30,15 @@ void	move_backward(t_player *p, t_level *lvl)
 
 void	move_left(t_player *p, t_level *lvl)
 {
-	double	delta;
-	double	dir_x;
-	double	dir_y;
+	float	delta;
+	float	dir_x;
+	float	dir_y;
 
 	delta = p->angle - PI_DIAGONAL;
 	if (delta < 0)
 		delta += PI2;
-	dir_x = -cos(delta);
-	dir_y = -sin(delta);
+	dir_x = -cosf(delta);
+	dir_y = -sinf(delta);
 	if (is_wall(lvl, (int)(p->x - (dir_x * (MOVE_SPEED + BOUNDARY))), \
 	(int)(p->y - (dir_y * (MOVE_SPEED + BOUNDARY)))))
 		return ;
@@ -48,15 +48,15 @@ void	move_left(t_player *p, t_level *lvl)
 
 void	move_right(t_player *p, t_level *lvl)
 {
-	double	delta;
-	double	dir_x;
-	double	dir_y;
+	float	delta;
+	float	dir_x;
+	float	dir_y;
 
 	delta = p->angle + PI_DIAGONAL;
 	if (delta > PI2)
 		delta -= PI2;
-	dir_x = cos(delta);
-	dir_y = sin(delta);
+	dir_x = cosf(delta);
+	dir_y = sinf(delta);
 	if (is_wall(lvl, (int)(p->x + (dir_x * (MOVE_SPEED + BOUNDARY))), \
 	(int)(p->y + (dir_y * (MOVE_SPEED + BOUNDARY)))))
 		return ;
